[css-typed-om] Generate whitelist from CSSProperties.json5.
Currently, any CSS property in CSSProperties.json5 with keywords will
implicitly accept keywords. This can cause crashes if the property
doesn't actually store keywords as CSSIdenifierValues.

We solved this by having a manual whitelist of properties we explicitly
support. However, this is annoying to maintain. So we generate this
whitelist from CSSProperties.json5 by explicitly specifying "Keyword"
as a typedom_type.

e.g. a property that doesn't have typedom_types will not be supported
by Typed OM. A property that has "Keyword" in typedom_types will
take CSSKeywordValues and reify as CSSKeywordValues.

We also add support for scroll-padding-* because the JSON for that
was already there and we just to write a test for it.

Diff of generated files:
https://gist.github.com/darrnshn/03cb617909abab7bf94aaaaad9874a15/revisions

Bug: 820299
Change-Id: I6fa01c3623161e0778c5f8764009c6e0cdd6bbf3